[refactor]: Nav에서 Dropdown 분리
