+++ frontend/src/components/Dropdown/dropdown.scss
@@ -1,0 +1,66 @@
+.dropdown {
+  position: relative;
+}
+
+.profile {
+  position: relative;
+  width: 50px;
+  height: 50px;
+}
+
+.profile__image {
+  cursor: pointer;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 50%;
+}
+
+.dropdown__content::before {
+  content: '';
+  position: absolute;
+  z-index: 0;
+  top: -5px;
+  right: 20px;
+  width: 10px;
+  height: 10px;
+  background: var(--dark);
+  transform: rotate(45deg);
+}
+.dropdown__content {
+  position: absolute;
+  top: 70px;
+  right: 0px;
+  width: 110px;
+  padding: 10px 10px;
+  background-color: var(--dark);
+  border-radius: 15px;
+  transition: 0.5s;
+  box-shadow: rgba(0, 0, 0, 0.15) 0px 3px 3px 0px;
+  visibility: hidden;
+  opacity: 0;
+}
+
+.dropdown__content--active {
+  visibility: visible;
+  opacity: 0.9;
+}
+
+.dropdown__item {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 15px;
+}
+
+.dropdown__link {
+  font-size: 15px;
+  font-weight: bold;
+  text-decoration: none;
+  color: var(--white);
+  cursor: pointer;
+
+  &:hover {
+    opacity: 0.3;
+  }
+}
