+++ frontend/src/components/LoginModal/loginModal.scss
@@ -1,3 +1,15 @@
+.login {
+  font-size: 15px;
+  font-weight: bold;
+  text-decoration: none;
+  color: var(--medium);
+  cursor: pointer;
+
+  &:hover {
+    opacity: 0.3;
+  }
+}
+
 .login-modal {
   position: fixed;
   top: 0;
@@ -47,7 +59,6 @@
     cursor: pointer;
   }
 }
-
 
 .login-modal__title {
   color: var(--black);
